+++ src/module.ts
@@ -1,10 +1,13 @@
 import { ModuleWithProviders, NgModule, Optional, SkipSelf } from '@angular/core';
-import { BrowserXhr, ConnectionBackend, Http as HttpAngular, ResponseOptions, XSRFStrategy } from '@angular/http';
+import { BrowserXhr, ConnectionBackend, RequestOptions, ResponseOptions, XSRFStrategy } from '@angular/http';
 import {
+  CancelRequestPlugin,
   DefaultOptionsToken,
   Events,
   Http,
   HttpEvents,
+  httpFactory,
+  HttpOverride,
   HttpPluginsToken,
   ParseResponsePlugin,
   ParseResponseToken,
@@ -35,7 +38,7 @@ export class HttpModule {
           useFactory: xhrBackendFactory,
           deps: [ BrowserXhr, ResponseOptions, XSRFStrategy, HttpEvents ]
         },
-        HttpAngular,
+        { provide: HttpOverride, useFactory: httpFactory, deps: [ ConnectionBackend, RequestOptions ] },
         { provide: ThrowExceptionStatusToken, useValue: null },
         { provide: ParseResponseToken, useClass: ThrowExceptionStatus, deps: [ ThrowExceptionStatusToken ], multi: true },
         { provide: Plugins, useClass: Plugins, deps: [ HttpPluginsToken ] },
@@ -56,6 +59,11 @@ export interface TypePlugins {
 }
 
 export const DefaultPlugins: any = [
+  {
+    provide: HttpPluginsToken,
+    useClass: CancelRequestPlugin,
+    multi: true
+  },
   {
     provide: HttpPluginsToken,
     useClass: ParseResponsePlugin,