+++ src/providers/index.ts
@@ -1,5 +1,6 @@
 export * from './backend/xhr_backend';
 export * from './http';
+export * from './http_override';
 export * from './options';
 export * from './plugins';
 export * from './mapper';