+++ src/providers/plugins/plugins.ts
@@ -12,8 +12,6 @@ export class Plugins {
 
   private plugins: Array<{[name: string]: Plugin}> = [];
 
-  private throwsException: boolean | Function = true;
-
   private options: Object = {};
 
   constructor(plugins?: Array<Plugin>) {
@@ -29,37 +27,6 @@ export class Plugins {
     }
 
     return this;
-  }
-
-  setThrowsException(throws: boolean | Function): this {
-    this.throwsException = throws;
-    return this;
-  }
-
-  getThrowsExceptionPlugin(plugin: string | Plugin): boolean | Function {
-
-    if (typeof plugin !== 'string') {
-      plugin = plugin.getName();
-    }
-
-    this.options[plugin] = this.options[plugin] || {};
-
-    let throwPlugin = 'throwsException' in this.options[plugin]
-      ? this.options[plugin].throwsException
-      : this.throwsException;
-
-    return throwPlugin;
-  }
-
-  isThrowsException(plugin: string | Plugin): boolean {
-
-    if (typeof plugin !== 'string') {
-      plugin = plugin.getName();
-    }
-
-    let throwPlugin = this.getThrowsExceptionPlugin(plugin);
-
-    return throwPlugin === true || typeof throwPlugin === 'function';
   }
 
   add(plugin: Plugin, priority?: number): this {
@@ -149,13 +116,15 @@ export class Plugins {
     }
   }
 
-  cleanOptions() {
+  cleanOptions(event: string) {
     this.forEach((plugin: Plugin) => {
       if ('restoreOptions' in plugin) {
-        plugin.restoreOptions();
+        if (event && event in plugin) {
+          plugin.restoreOptions();
+        }
       }
+      this.options[plugin.getName()] = {};
     });
-    this.options = {};
   }
 
   setOptions(options: Object): this {
@@ -166,6 +135,8 @@ export class Plugins {
         if ('setOptions' in plugin) {
           plugin.setOptions(options[AllPlugins]);
         }
+
+        this.options[plugin.getName()] = options[AllPlugins];
       });
 
       delete options[AllPlugins];
@@ -182,6 +153,8 @@ export class Plugins {
       if ('setOptions' in plugin) {
         plugin.setOptions(options[pluginName]);
       }
+
+      this.options[plugin.getName()] = options[pluginName];
     }
 
     this.options = options;
@@ -193,7 +166,7 @@ export class Plugins {
       throw new Error(`Event '${event}' not exists`);
     }
 
-    this.forEach((plugin: Plugin) => {
+    this.forEach((plugin: any) => {
       if (!(event in plugin)) {
         return;
       }
@@ -203,8 +176,8 @@ export class Plugins {
         method.apply(plugin, params);
       } catch (ex) {
 
-        if (this.isThrowsException(plugin)) {
-          let callbackException = this.getThrowsExceptionPlugin(plugin);
+        if (plugin.getThrowsException && plugin.getThrowsException()) {
+          let callbackException = plugin.getThrowsException();
           if (typeof callbackException === 'function') {
             callbackException(ex);
             return;